Minecaching v0.2.2.2: Notification & Localization Update
- Added ArchiveCacheCommand and DisableCacheCommand
- Renamed VerifyCacheCommand to PublishCacheCommand
- MinecacheStatus#VERIFIED has been renamed to PUBLISHED, and the id from ACTIVE to PUBLISHED
- Updated PublishCacheCommand
- 2 new events, MinecacheArchivedEvent and MinecacheDisabledEvent
- MinecacheVerifiedEvent has been renamed to MinecachePublishedEvent
- Update @since versions in all events
- Renamed MCPluginMessages to MCMessages
- Added MinecachingAPI#archiveMinecache and MinecachingAPI#disableMinecache
- Renamed MinecachingAPI#verifyMinecache to MinecachingAPI#publishMinecache

Took 1 hour 13 minutes

+++ src/main/java/net/realdarkstudios/minecaching/api/PlayerDataObject.java
@@ -206,13 +206,21 @@ public class PlayerDataObject {
             }
         } else if (Config.getInstance().getPlayerDataVersion() != MinecachingAPI.getPlayerDataVersion()) {
             try {
+                if (!plrFile.canWrite()) throw new Exception();
+                PlayerDataObject pdo = getPDO(uuid, yaml, plrFile, useEmptyMinecache);
                 plrFile.delete();
                 plrFile.createNewFile();
+                pdo.saveData();
+                return pdo;
             } catch (Exception e) {
                 MinecachingAPI.tWarning("error.plugin.updatefile", uuid + ".yml");
             }
         }
 
+        return new PlayerDataObject(uuid, yaml, plrFile, useEmptyMinecache);
+    }
+
+    private static PlayerDataObject getPDO(UUID uuid, YamlConfiguration yaml, File plrFile, boolean useEmptyMinecache) {
         return new PlayerDataObject(uuid, yaml, plrFile, useEmptyMinecache);
     }
 }
